+++ yesod-core/test/RouteLeafCompile/MissingEndpoint.hs
@@ -3,5 +3,5 @@ module MissingEndpoint where
 import Yesod.Core.RouteLeaf
 import YesodCoreTest.RouteLeaf.Foundation
 
-bad :: RouteLeaf AccountR -> Bool
+bad :: RouteLeaves AccountR -> Bool
 bad (LeafItemR _) = True