+++ yesod-core/test/RouteLeafCompile/NestedPattern.hs
@@ -3,5 +3,5 @@ module NestedPattern where
 import Yesod.Core.RouteLeaf
 import YesodCoreTest.RouteLeaf.Foundation
 
-bad :: RouteLeaf OrgR -> Bool
+bad :: RouteLeaves OrgR -> Bool
 bad (DelegationR _) = True